moved spec.opacity.xsection to spec.lpf.auto_xsection
