+++ tests/figures/xs/mdbhit_extract.py
@@ -1,4 +1,4 @@
-from exojax.spec import xsection
+from exojax.spec.lpf import auto_xsection
 from exojax.spec.hitran import SijT, doppler_sigma, gamma_hitran, gamma_natural
 from exojax.spec.exomol import gamma_exomol
 from exojax.spec import moldb
