+++ tests/integration/moldb/comparison_api_xs.py
@@ -5,7 +5,7 @@ from jax.config import config
 from exojax.spec.hitran import normalized_doppler_sigma
 from exojax.spec.set_ditgrid import ditgrid_log_interval
 from exojax.spec.hitran import SijT, doppler_sigma, gamma_hitran, gamma_natural
-from exojax.spec import xsection as lpf_xsection
+from exojax.spec.lpf import auto_xsection as lpf_xsection
 from exojax.spec import initspec
 from exojax.spec.modit import xsvector as modit_xsvector
 from exojax.spec.lpf import xsvector as lpf_xsvector
